refactor: separate BackgroundProducerWorker code into files
